Adding changes in ClimbingSteps with three solutions

+++ Leetcode/src/main/java/easy/BinaryAdd.java
@@ -9,28 +9,32 @@ public class BinaryAdd {
 // add binary numbers and return characters
 	
 	public static void main(String[] args) throws InterruptedException {
-		long st = System.currentTimeMillis();
-	//	System.out.println(addBinaryNum("1101", "0111"));
+		BigInteger st = System.currentTimeMillis();
+		// Worked but not accepted.
+		//System.out.println(addBinaryNum("1101", "0111"));
 	    System.out.println(addBinary("1101","0111"));
+	    
+	    //Testing
 	    //Thread.sleep(10);
-	    long et = System.currentTimeMillis();
+	    
+	    BigInteger et = System.currentTimeMillis();
 	    System.out.println((et-st));
 	}
 	  static String addBinaryNum(String a, String b){
 		  return num2Bin(bin2Num(a)+bin2Num(b));
 	  }
-	  static int bin2Num(String bin){
-		 int unressolved = Integer.parseInt(bin);
-		 int resolved=0;
-		 int count=0;
+	  static BigInteger bin2Num(String bin){
+		 BigInteger unressolved = Integer.parseInt(bin);
+		 BigInteger resolved=0;
+		 BigInteger count=0;
 		 while(unressolved>0){
-			resolved = (int) (resolved+((unressolved%2)*Math.pow(2, count))); 
+			resolved = (BigInteger) (resolved+((unressolved%2)*Math.pow(2, count))); 
 			count++;
 			unressolved= unressolved/10;
 		 }
 		 return resolved;
 	}
-	  static String num2Bin(int num){
+	  static String num2Bin(BigInteger num){
 		 String i="";
 		  while(num>1){
 			  i=i+num%2;
@@ -45,9 +49,9 @@ public class BinaryAdd {
 	  static String addBinary(String a, String b){
 	  
 	  StringBuilder sb = new StringBuilder();	  
-	  int carry =0 ;
-	  int i = a.length()-1;
-	  int j = b.length()-1;
+	  BigInteger carry =0 ;
+	  BigInteger i = a.length()-1;
+	  BigInteger j = b.length()-1;
 	  while(i>=0|| j>=0 || carry==1){
 		if(i>=0){
 			carry += a.charAt(i--)-'0';
